requirements.txt
langchain
langchain_experimental
openai
wikipedia
requests
beautifulsoup4
faiss-cpu
duckduckgo-search
html2text
validators
tiktoken